requirements.txt
albucore==0.0.24
albumentations==2.0.8
altair==5.5.0
annotated-types==0.7.0
atomicwrites==1.4.0
blinker==1.9.0
cachetools==5.5.2
contourpy==1.3.2
cycler==0.12.1
filelock==3.18.0
fonttools==4.58.2
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
hf-xet==1.1.3
huggingface-hub==0.32.4
ipywidgets==8.1.7
joblib==1.5.1
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
matplotlib==3.10.3
mpmath==1.3.0
narwhals==1.42.0
networkx==3.5
numpy==2.3.0
onnx==1.18.0
opencv-python-headless==4.11.0.86
pandas==2.3.0
pillow==11.2.1
protobuf==6.31.1
pyarrow==20.0.0
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
pyls-spyder==0.4.0
pyparsing==3.2.3
PyQt5==5.15.10
PyQtWebEngine==5.15.6
pytz==2025.2
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
seaborn==0.13.2
segmentation_models_pytorch==0.5.0
setuptools==78.1.1
simsimd==6.4.9
smmap==5.0.2
streamlit==1.45.1
stringzilla==3.12.5
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
timm==1.0.15
toml==0.10.2
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
triton==3.3.1
typing-inspection==0.4.1
tzdata==2025.2
wheel==0.45.1
widgetsnbextension==4.0.14
xgboost==3.0.2
